pyproject.toml
[tool.poetry]
name = "smart-pantry"
version = "0.1.0"
description = ""
authors = ["Patrick <patrick.pwall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.8"
fastapi = "^0.61.1"
tortoise-orm = "^0.16.15"
asyncpg = "^0.21.0"
pytest-cov = "^2.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
aiosqlite = "^0.15.0"
faker = "^4.1.3"
pytest-asyncio = "^0.14.0"
asynctest = "^0.13.0"
pytest-mock = "^3.3.1"
requests = "^2.24.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
